KNN 2
By leveraging the k-Nearest Neighbors (kNN) algorithm, we developed a highly accurate predictive model to classify cars based on various features such as buying price, maintenance cost, number of doors, capacity, luggage boot size, and safety rating. The model was trained on a dataset with these attributes and successfully achieved a remarkable 94% accuracy rate. This high level of accuracy signifies the model's strong ability to predict the classification of cars (unacceptable, acceptable, good, or very good) based on the given features. The kNN algorithm, known for its simplicity and effectiveness in classification tasks, was instrumental in achieving this performance, demonstrating its applicability to real-world datasets in the automotive industry.